requirements.txt
# Dependencies

streamlit>=1.20
google-genai>=0.3.0
python-dotenv>=1.0.0
email-validator>=1.1.3

